백엔드: 교재, 교재문제, 독서책, 학생로그, 카테고리 기능 추가 교재: 교재 및 교재문제 등록, 수정, 제거 기능 독서책: 학생이 읽을 도서 엔티티 학생로그: Study 엔티티 코스에 맞게 오늘자 해야할 수업 코스 가져오기 기능 추가, 코스 변경시 해당 코스에 맞는 새로운 로그 생성.

+++ src/main/java/SNP/management/domain/repository/CategoryDataJpa.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CategoryDataJpa extends JpaRepository<Category, Long> {
+public interface CategoryDataJpa extends JpaRepository<Category, Integer> {
 
     public Optional<Category> findByName(String id);
 }
